Fixed races/gliches in comp_merge and the input generation
There is a lot in this change, but the most important was a deep look
at the merge component which had several issues.  The book version is
either misleading or wrong.

The way comp_contr is also totally broken but that's on me.  I replace
with pure token flow circuit that works.